remove admin mobile portal
