+++ src/App/Header/MobileNav/index.ts
@@ -1,2 +1,1 @@
-export * from "./Menu/AdminMobileNavPortal";
 export * from "./Opener";
