+++ src/pages/Admin/Charity/Nav.tsx
@@ -4,7 +4,7 @@ import { adminRoutes } from "constants/routes";
 
 export default function Nav() {
   return (
-    <div className="hidden lg:flex justify-end">
+    <div className="flex justify-end">
       <NavLink end to={adminRoutes.index} className={styler}>
         Dashboard
       </NavLink>
