+++ src/slices/components/index.ts
@@ -1,8 +1,6 @@
 import { combineReducers } from "redux";
 import marketFilter from "./marketFilter";
-import mobileNav from "./mobileNav";
 
 export const componentReducer = combineReducers({
-  mobileNav,
   marketFilter,
 });
